refactor: implement EventEmitter class and integrate with RenderManager and ZoomManager
