+++ packages/pdf-core/src/utils/index.ts
@@ -1,3 +1,5 @@
+export { EventEmitter } from './EventEmitter';
+
 /**
  * Check if we're running in a browser environment
  */
